Zip file contains the whole code and story board for E-Airlines iOS application. Also attached a video where you can see the execution in iphone 11 pro max simulator. Moreover attached a PPT where you can find MVC diagram and output samples.
In this iOS application, the user can search for the status of flights that are currently running. Also, they can search for all the available flights from a location to another on a particular date. Moreover, this application can fetch all the nearby airports upon entering a location. All this data will be a real-time one which is fetched from Lufthansa Airlines API.
Note: You need to have your own API token/key to implement this type of project. Which you can get by contacting API providers.